requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
idna==2.10
kiwisolver==1.3.0
matplotlib==3.3.2
numpy==1.19.3
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.11
youtube-dl==2020.11.12
mutagen~=1.45.1
pydub~=0.24.1